Initial

+++ day06/ex00/srcs/ft_strncpy.c
@@ -1,7 +1,7 @@
 char *ft_strncpy(char *dest, char *src, unsigned int n)
 {
     int i = 0;
-    while (i != n && src[i] != '\0')
+    while (i != (int)n && src[i] != '\0')
     {
         dest[i] = src[i];
         i++;
